Add subnets to the host and populate it when scanning a host (#1344)
This PR adds a `subnets` column to the `hosts` table. This in an attempt
to get rid of the redundant IP churn we saw on `arequipa`. This is under
the assumption that the churn is related to the way we deal with
intermittent dns resolve issues, which up until now we've treated as
redundant IPs because the `ipFilter` returns a simple yes/no bool.

Fixes https://github.com/SiaFoundation/renterd/issues/1342

---------

Co-authored-by: Christopher Schinnerl <chris@sia.tech>