Add contract pruning to the bus (#1472)
This PR moves contract pruning over to the bus. I kept the
`/rhp/contract/:id/roots` endpoint on the worker. This PR also uses a
temporary table to store the roots we fetch from the host, this allows
us to calculate which indices are prunable without fetching the contract
roots from the database. This saves on network traffic but also turned
out to make contract pruning about 2-3 times as fast. Not sure if this
performance boost is linear though because the slow part might be the
deletions.. pruning ~200MiB from 1.5 TiB contracts takes ~11s on average
as opposed to ~25s when we fetch roots and do the diff manually.