Remove `directories` table (#1665)
We recently (hot)fixed a bug when renaming directories. Unfortunately I
overlooked the case where a directory can be moved, wasn't even aware
our UI allowed dragging-and-dropping folders on the `..` directory like
that. This RP fixes that by removing the `directories` table entirely,
rewriting objects and rename queries from scratch.

@ChrisSchinnerl I tested this on both SQLite and MySQL, both the
behaviour in the UI as well as the migration. It's a single commit so we
can easily cherry-pick it on `master` should we decide to roll this out
as another hotfix. I asked Alex to deploy this branch and play test the
UI as well. The "All Files" view sometimes confuses me but since all of
our tests pass I decided not to tweak the logic.

**edit: play tested by both Alex and myself and it behaves as expected
now, still slightly uneasy about rewriting all the queries but all tests
pass**

Fixes https://github.com/SiaFoundation/renterd/issues/1599

Co-authored-by: Christopher Schinnerl <chris@sia.tech>